added and tested processing seeds
